# Improve performance of `f(*a, **kw)` style method dispatch