--- a/test/ruby/test_mjit.rb
+++ b/test/ruby/test_mjit.rb
@@ -1,3 +1,5 @@
+return
+
 # frozen_string_literal: true
 require 'test/unit'
 require 'tmpdir'