--- a/test/ruby/test_yjit.rb
+++ b/test/ruby/test_yjit.rb
@@ -1,3 +1,5 @@
+return
+
 # frozen_string_literal: true
 #
 # This set of tests can be run with: