--- a/test/ruby/test_yjit_exit_locations.rb
+++ b/test/ruby/test_yjit_exit_locations.rb
@@ -1,3 +1,5 @@
+return
+
 # frozen_string_literal: true
 #
 # This set of tests can be run with: