@ko1 2022-12-13
```ruby
# benchmark code

require 'forwardable'
require 'pp'
require 'benchmark/ips'

class Obj
  extend Forwardable

  attr_accessor :other

  def initialize
    @other = Other.new
  end

  def foo_without_splat
    @other.foo
  end

  def foo_with_splat_with_name(*args)
    @other.foo(*args)
  end

  def foo_with_splat_and_double_splat_with_name(*args, **kw)
    @other.foo(*args, **kw)
  end

  begin
    binding.eval <<~RUBY
      def foo_with_splat(*)
        @other.foo(*)
      end

      def foo_with_splat_and_double_splat(*, **)
        @other.foo(*, **)
      end
    RUBY
  rescue SyntaxError
    # ignore error on Ruby 3.1
  end

  def foo_with_triple_dots(...)
    @other.foo(...)
  end

  delegate :foo => :@other
end

class Other
  def foo() end
end

o = Obj.new

Benchmark.ips do |x|
  x.report 'foo()' do
    o.other.foo
  end

  x.report 'dele()' do
    o.foo_without_splat
  end

  x.report 'dele(*)' do
    o.foo_with_splat
  end if o.respond_to? :foo_with_splat

  x.report 'dele(*a)' do
    o.foo_with_splat_with_name
  end

  x.report 'dele(*a, **kw)' do
    o.foo_with_splat_and_double_splat_with_name
  end

  x.report 'dele(*, **)' do
    o.foo_with_splat_and_double_splat
  end if o.respond_to? :foo_with_splat_and_double_splat

  x.report 'dele(...)' do
    o.foo_with_triple_dots
  end

  x.report 'forwardable' do
    o.foo
  end
end
```