@k0kubun 2022-12-13
Have you run non-microbenchmarks? While JIT support is not ready, you could still compare the performance of interpreters on [yjit-bench](https://github.com/shopify/yjit-bench) with `./run_benchmarks.rb -e "before::/path/to/ruby1" -e "after::/path/to/ruby2"`.