@maximecb 2022-12-13
Hi Koichi. I think it's good that you found a way to avoid some unnecessary allocations in splat calls.

I'm a little bit concerned though because we're just 12 days away from the 3.2 release date. That limits the time we have to implement this in YJIT and to test that everything is 100% stable. Merging this PR without YJIT support could impact YJIT's performance, and some of us may be on vacation next week and unavailable.

Just some words of caution. We will do our best to help you get this working in YJIT, but we may want to delay merging until after the 3.2 release depending.