@k0kubun 2022-12-13
> Have you run non-microbenchmarks?

I just did that myself. I ran the "headline" benchmarks (the most real-world set of benchmarks) of yjit-bench:

```
before: ruby 3.2.0dev (2022-12-13T16:00:22Z master a66a69865d) [x86_64-linux]
after: ruby 3.2.0dev (2022-12-13T18:31:27Z opt_kwrest 5403761a81) [x86_64-linux]

-------------  -----------  ----------  ----------  ----------  ------------  -------------
bench          before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
activerecord   105.4        0.4         101.7       0.4         1.04          1.04
hexapdf        1802.1       0.5         1816.2      0.3         0.99          0.99
liquid-render  113.9        0.7         114.9       0.6         0.99          1.00
mail           96.9         0.2         96.0        0.2         1.01          1.00
psych-load     1389.5       0.1         1365.3      0.2         1.02          1.02
railsbench     1563.4       0.8         1575.7      1.3         0.99          1.00
ruby-lsp       48.1         11.2        48.7        12.8        0.99          1.03
-------------  -----------  ----------  ----------  ----------  ------------  -------------
```

https://gist.github.com/k0kubun/2a2cfd2af4b2ec8b9b46ef605f8d3b99