@ko1 2022-12-13
> I just did that myself. I ran the "headline" benchmarks (the most real-world set of benchmarks) of yjit-bench:

Thank you. ActiveRecord seems to use this kind of delegators.