@ko1 2022-12-13
> Just some words of caution. We will do our best to help you get this working in YJIT, but we may want to delay merging until after the 3.2 release depending.

Yes. I noted about the deadline of Ruby 3.2 on the original ticket: https://bugs.ruby-lang.org/issues/19165#note-5
And I think it is difficult to merge it on Ruby 3.2, too.

Known issues, the following needs catch up:

* YJIT (test/ruby/test_yjit* works, but it seems other bugs)
* MJIT
* Typeprof