@ko1 2022-12-15
I heard typeprof had been fixed for this patch.