@ko1 2022-12-15
On the devmeeting, this patch will be merged on Ruby 3.3, not on Ruby 3.2.
Let's keep discussion.